Add types for schema
`RootDefinition` is the outer most type that picks up `properties`

`FirstLevelDefinitions` is everything inside of `properties: {...}`

`SchemaDefinition` is a recursive union over all possible rule definitions. A rule definition is any property that does not exist as a property in `AllTypes`. Those could be considered 'reserved'.

Depending on the type specified, like `{ type: 'string' }`, the correct type hints are available for the applicable formatters/validators/modifiers.

The `required` property should now validate that the provided value(s) actually are defined.

Notes:

`required` and `additionalProperties` are moved from the same level as `properties` to the same level as the definitions. This is only changed for the topmost rules modifiers. The rest of them stays on the same level as before.

The types are not really ergonomic. They work because they are inferred from the function argument. As such, using them with a separately defined schema might be cumbersome. Maybe with a `satisfies` or something.

They do not discriminate well, so a few `!` assertions had to put in there.

---